+++ example/test/widget_test.dart
@@ -6,7 +6,6 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 
-import 'package:flutter_test/flutter_test.dart';