+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+## 1.1.2
+
+* News Validations implemented isPDF, isTxt, isChm, isVector, isHTML
+
 ## 1.1.0
 
 * Changed name in  Function removeCaracteres to removeCharacters
 * this can break your code please check before update.
-* more formar in validation photos and videos add 
-* more validations isRG
+* more formar in validation photos and videos 
+* add more validations isRG
 
 ## 1.0.1